+++ controllers/product.controller.js
@@ -24,21 +24,24 @@ exports.getRegisterProduct = async (request, response) => {
 // POST /product: register a new product
 exports.postRegisterProduct = async (request, response) => {
     try {
+        const imagen = request.file ? request.file.path : request.body.image;
+
         const product = new Product (
             request.body.idTaller,
             request.body.Nombre,
             request.body.PrecioUnitario,
             request.body.idCategoria,
             request.body.Descripcion,
-            request.body.imagen,
+            imagen,
             request.body.Disponible,
         );
 
         // Success
-        await Product.save();
+        await product.save();
         response.status(201).json({ message: "Product added successfully" });
     } catch (err) {
         // Error
+        console.log(err);
         response.status(500).json({ message: "Failed to add a product", err });
     }
 }
@@ -70,19 +73,18 @@ exports.getOneProduct = async (request, response) => {
 
 // PUT: modify an excisting product
 exports.updateProduct = async (request, response) => {
-    try {        
+    try { 
         const id = request.params.idProduct;
         
-        const idTaller = await Workshops.getId(request.body.nombreTaller); 
-        const idCategoria = await Category.getId(request.body.Categoria);
+        const imagen = request.file ? request.file.path : request.body.image;
         
         const productData = {
-            idTaller: idTaller,
+            idTaller: request.body.idTaller,
             Nombre: request.body.Nombre,
             PrecioUnitario: request.body.PrecioUnitario,
-            idCategoria: idCategoria,
+            idCategoria: request.body.idCategoria,
             Descripcion: request.body.Descripcion,
-            imagen: request.body.imagen,
+            imagen: imagen,
             Disponible: request.body.Disponible,
         };
 
@@ -92,6 +94,7 @@ exports.updateProduct = async (request, response) => {
         response.status(200).json({ message: "Product modified successfully" }); 
     } catch (err) {
         // Error
+        console.log(err);
         response.status(500).json({ message: "Failed to modify a product", err });
     }
 }