+++ routes/product.routes.js
@@ -1,11 +1,19 @@
 const express = require("express");
 const router = express.Router();
+const multer = require("multer");
 
 const productController = require("../controllers/product.controller");
 
+const storage = multer.diskStorage({
+    destination: (request, file, cb) => cb(null, "uploads/"),
+    filename: (request, file, cb) => cb(null, Date.now() + "_" + file.originalname)
+})
+
+const upload = multer({ storage });
+
 // POST: Register a new product
 router.get("/add", productController.getRegisterProduct);
-router.post("/add", productController.postRegisterProduct);
+router.post("/add", upload.single("image"), productController.postRegisterProduct);
 
 // GET: Search products
 router.get('/search', productController.searchP);
@@ -30,7 +38,7 @@ router.delete("/:idProduct", productController.deleteProduct);
 
 //GET: Modify product
 router.get("/:idProduct/update", productController.getOneProduct);
-router.put("/:idProduct/update", productController.updateProduct);
+router.put("/:idProduct/update", upload.single("image"), productController.updateProduct);
 
 
 module.exports = router;