+++ routes/qr.routes.js
@@ -3,7 +3,7 @@ const router = express.Router();
 
 const qrController = require("../controllers/qr.controller.js");
 
-router.get("/create", qrController.createQR);
+router.post("/create", qrController.createQR);
 router.post("/validate", qrController.validateQR);
 
 module.exports = router;