+++ routes/supplyBatch.routes.js
@@ -26,10 +26,10 @@ router.get("/:idInsumo", supplyBatchController.getOneSupplyBatch);
 router.post("/addBatch", supplyBatchController.addSupply);
 
 /**
- * DELETE /:idInventario
- * Deletes a specific supply batch by its inventory ID.
+ * DELETE
+ * Deletes a specific supply batch by its expiration date.
  */
-router.delete("/:idInventario", supplyBatchController.deleteSupplyBatch);
+router.post("/delete", supplyBatchController.deleteSupplyBatch);
 
 // Define a GET endpoint to retrieve supply batches expiration date and acquisition type
 router.get("/filter/data", supplyBatchController.getFilterData);
@@ -42,6 +42,13 @@ router.get("/acquisition/types", supplyBatchController.getAcquisitionTypes);
 
 // Define a POST endpoint to modify one supply batch
 router.post("/update/:idSupplyBatch", supplyBatchController.modifySupplyBatch);
+
+// Define a GET endpoint to retrieve one supply batch for modify
+router.get("/modify/:idSupplyBatch", supplyBatchController.getSupplyBatchOne);
+
+// Define a GET endpoint to retrieve supply batches by date and supply id
+// Example: /dates/22-11-2025/7ec9e58...
+router.get("/dates/:date/:idInsumo", supplyBatchController.getSupplyBatchDates);
 
 // Export the router to be used in the main application
 module.exports = router;